Use PLMN instead of SPN override for Rogers, DTAG and Telstra
Rogers, DTAG and Telstra want the actual SIM card SPN to be displayed

Having SPN override in /etc/spn-conf.xml always overrides SPN of the SIM
card irrespective of Wi-Fi calling registered or not. Carriers doesn't
like this as they don't see SIM card SPN even when the device is camped
on cellular RATs.

This CL
 1) Removes the carriers Rogers, DTAG and Telstra from spn-conf.xml.
 2) Adds a logic in updateSpnDisplay() to use PLMN if SPN is empty
    and Voice over Wi-Fi is registered case.

Test: Tested using T-Mobile SIM card after removing T-Mobile entries
      from the /etc/spn-conf.xml.

Bug: 64137294
Bug: 64139429
Change-Id: Ia350b4d06739f47d5da4eba0a5b0970af6db5f36