Support auto-downgrade of video calls when mobile data is metered.
The previous code supported downgrading video calls (either to audio,
or pausing video, or terminating the call, depending on carrier reqs)
when mobile data is disabled.

These changes provide support for downgrading a video call over wifi
when it hands over to LTE (that work is in onCallHandover).  The rest
of the changes in ImsPhoneCallTracker were to refactor the downgrade
code so it could be reused.

In ImsPhoneConnection, added support for temporarily disabling VT on the
local side.  This way, if a wifi video call downgrades to audio-only due
to mobile data being disabled, the "upgrade to video" button will not
be available allowing the user to upgrade back to video (thereby defeating
the fact data is disabled).

Test: Manual
Bug: 36900451
Change-Id: Ib29c7b2fb31f4dcf505609bedfbaca0be7f8f265
Merged-In: I07eb69d83a727ddd21992832753d8ecc187dbfd7