fix TelephonyNetworkFactory unit test failure
there might be a race condition: the main testing thread executes
testcase (after setup) before the looper of the handlerThread is ready,
or before the message has been successfully delivered to the handerThread.
As a result, the main thread is creating a TelephonyNetworkFactory
object with a null DCT object.
the fix is create DcT object when the looper is ready and wait till DCT is
created successfully.

Bug: 29579544
Test: runtest --path
Change-Id: I9d00c78fc3290d2a0ec2e493c5a13879598005af