Fix timezone cannot be updated automatically when no NITZ received
When no NITZ received, timezone can be updated by mobile country code.
This procedure is in pollStateDone method and only happens when network 
changed. So if user manually changed timezone then turn on AutoTimezone 
Setting and network is stable, timezone cannot be automatically updated.

To fix, when turn on auto timezone setting, if there is no NITZ timezone
, immediately update timezone by country code.

Test: Build pass