Adding parcelable class checking for call extras.
Currently, call extras originate from vendor code in the ImsCallProfile.
It is possible for a vendor to put in a package-local parcelable class,
which can cause problems in Telecom and other consumers of the
parcelable class.  The result is that the Bundle and its contents will be
lost.
This change ensures that the Connection class, which is currently the
primary entry point from extras in the ImsCallProfile class will filter
out Bundle contents which are not valid.  The exisint Bundle.filterValues
method accomplishes this by ensuring only primitive types and android.*
namespace parcelables are included in the Bundle.

Although ImsCallProfile is technically the best place to enforce this,
the vendor modem code currently runs in the same process as Telephony, so
the parceling code there never gets run.  Hence the filtering here.

Also, added a new unit test for the ImsCallProfile to verify the parcel
enforcement happening at that level as well.

Test: Manual, also wrote new unit tests
Bug: 65562929
Change-Id: Iabaae8d0a44a1114b3618937e4238e9f4619cf5a