Fix issue where video call continues on LTE when wifi lost and data off.
Due to changes in b/36900451 to decouple VT from the data enabled state
of the device, there was a regression in the case where WIFI is lost during
a WIFI video call and mobile data is off.

Changed code to properly handle this case by triggering a disconnect when
video pause signalling is in use.

Test: Manual - turn off wifi when on WIFI video call and mobile data is
disabled.  Regression tested mobile data off auto-pausing.
Bug: 64531514

Change-Id: I6561f653cd446b74f127951e02d482f7c9267c9d