Don't cancel CS reject cause notifications
Notification should stay until the user dismisses it. Since the error
that causes this notification is a permanent error (until power cycle or
SIM swap), we shouldn't cancel it when we enter a new service state or
the SIM is removed.

Test: manual
Merged-In: I25735ca2c13a655a604d76880a805d9cf63ad1f6
Change-Id: I25735ca2c13a655a604d76880a805d9cf63ad1f6
Fixes: 64194949