Use Android data for MCC-based time zone lookup
Instead of using IANA's time zone ordering from zone.tab,
use Android data for MCC-based time zone lookup.

This allows us to remove a code hack for Australia (
which was forcing the use of Australia/Sydney over the
Australia/Lord_Howe given in the IANA data) and
improves a few others. It also means that the selection
is controlled in files mastered by Android and not files
taken directly from IANA.

With 2017b rules data it means the following
differences from prior behavior:

[ISO code]: Old answer, New answer.

[fm]:  Old: Pacific/Chuuk, New: Pacific/Pohnpei
[ru]:  Old: Europe/Kaliningrad, New: Europe/Moscow
[um]:  Old: Pacific/Midway, New: Pacific/Wake
[uz]:  Old: Asia/Samarkand, New: Asia/Tashkent

In all cases these are the zone IDs understood to cover the
region's capital.

This change only affects time zone lookups when only
the MCC is known (i.e. in the absence of NITZ).

Bug: 36971201
Bug: 62418027
Test: All known country ISO codes (and some unknown) tried
Change-Id: I7ee2d133802d2d91f36b247b05974c77abb957dc