Remove empty override of saveClirSetting in ImsPhoneBase.
When the CLIR setting is changed via GSM, the base Phone#saveClirSetting
method gets called to save the user's CLIR setting to shared prefs.
It turns out that when the IMS code was written, an empty stub override
was added for ImsPhone.  If an attempt was made to save the CLIR setting
over IMS, the network would be updated but the shared prefs would not be.

Removing the empty override ensures the default base Phone version of the
save logic gets used.

Test: Manual, performed CLIR set operations over IMS and verified saves.
Bug: 68029544
Change-Id: I678fccd8ec7c610002b503a18261e3807586af4d