Add some delays for the carrier privilege rules loading.
When UiccProfile is created or updated, the CarrierPriviligeRules might
be reloaded asynchronously. When the loading is complete, a message will
be sent out for notification. If the test finishes earlier than that,
the message will be sent to a dead thread. This won't break anything but
it is still wothy to fix it by adding some delays to the tests.

Bug: 64131518
Test: Basic telephony sanity
Change-Id: I18a3c30c300355601a66b553d8fe571ac5ad7492