The re-try flag is not set to true when re-sending SMS at second time
Retry flag for sending SMS message is set when send IMS SMS.
If this flag is true, the device re-sends SMS over SGs and if false,
the device re-sends SMS over IMS.
The problem is that the flag is set to true only first time.
So, if second retry fail occurs, the device re-sends SMS over IMS.
But it should send SMS over SGs.

To solve this issue, always set ImsSmsMessage#retry to true
if the number of retransmissions is 1 or more.

Test: auto - Confirmed that run RILTest and it has no problem.
Bug: 69871520
Change-Id: I395a670af6f69eb830b0c2ed7c92b8469012997b