Switch MccTable over to using TimeServiceHelper
Switch MccTable over to using TimeServiceHelper; there's
some logic that should be shared or identical for detecting
if the device automatically defaulted the time zone to GMT
on boot. This will slightly alter ServiceStateTracker
behavior (the other user of TimeServiceHelper), which
previously did not check for GMT. This new
behavior looks more correct given bug 64056758.

This commit also switches MccTable over to using
TimeServiceHelper to set the device time zone.

This alters behavior of MccTable and means there will be
an extra broadcast of ACTION_NETWORK_SET_TIMEZONE if MccTable
sets the time zone (which only happens if the device has
no time zone set). Although this MccTable path is not strictly
an NITZ update it's still a "telephony" update so broadly
qualifies as a "network time zone update".

MccTable should probably also call
TimeServiceHelper.isTimeZoneDetectionEnabled() before setting
the device time zone but that change is not being made as
part of this commit.

Test: atest FrameworksTelephonyTests
Bug: 64056758
Bug: 63743683
Change-Id: I1c17dab0c93b5827f33eb34d472453d048880bb7