Support the alpha identifier of REFRESH command
According to the clause 6.4.7 of ETSI TS 102 223, the terminal shall use
the alpha identifier to inform the user if it is provided in the REFRESH
command and is not a null data object. The idle mode text shall also be
removed when the terminal receives the REFRESH command with "reset" or
"initialization", but there is another trigger to do it. Please refer to
the broadcast "com.android.internal.stk.icc_status_change".

Bug: 70370632
Test: Confirmed the expected behavior and no degradation.

Change-Id: I76534a791efe371f8265c7a24bb6ddeead58ec47