Time Detection improvements
This change mostly modifies time detection logic to be
clearer, safer around wake locks. It adds / de-dupes
logging for improved debugging.

It also modifies mNitzUpdatedTime:

mNitzUpdatedTime was added to GsmServiceStateTracker in
commit a27421a306c49fbe9b3823b30f7ab1cd58b28854
and is used to track whether NITZ has been used. The original
purpose was to determine whether it was appropriate to try
to set the time zone using the network MCC (it's not if NITZ
has been received). The logic around this seems generally
bogus and the flag state was not maintained historically in
various paths (e.g. fixTimeZone()). This change renames
the field to reflect actual usage and moves it to be with
the Time Zone detection code. Later changes will improve the
time zone detection logic.

Bug: 63743683
Test: atest FrameworksTelephonyTests
Change-Id: Ic0925b1096eb20e35fb284100b54e82fc980839f