Remove unused system property setting code
Remove unused "gsm.nitz.time" system property setting code. This
system property dates from < 2008.

This system property is set as a side effect from NitzStateMachine
when NITZ data is received and not throttled.

The property is set even when the NITZ time signal is not used
to set the device system clock and no other information from NITZ
is recorded in this way.

No usages of this system property can be found in code. It appears
as though it was once used from radio stress tests and bugs suggest
it has occasionally been referenced but it is not very useful on its
own: there's more useful information in the (volatile) dumpsys /
bugreport logs.

Bug: 63743683
Test: atest FrameworksTelephonyTests
Change-Id: Ib077421d9a7c7f542176815c1f1fe1ce4e9f6113