Make state transitions clearer in NitzStateMachine
All state transitions are triggered in "handleXyz"
methods and ServiceStateTracker calls them at obvious
places.

Some of the logic when the MCC becomes known has been
pulled together into a single method from several
separate calls. This change may introduce some
functional changes as a result of these changes.
A field has been renamed to try to make the logic
still clearer now the state is not inspected by
ServiceStateTracker.

An additional call has been added to
handleNetworkUnavailable() in a place where it seems
appropriate.

The interface between ServiceStateTracker and
NitzStateMachine should be stable at this point,
allowing for NitzStateMachine to become an abstract
or interface class and enable an alternative
implementation to be swapped in.

Test: atest FrameworksTelephonyTests
Bug: 63743683
Change-Id: I2e03f9bbce06e2a03b32e467deb2a0a269adecb8