Disable Flaky IccCardProxy Tests
Mockito is occasionally returning strange errors in these
tests that appear to be due to some multithreading issue.
I have followed the recommended steps, but I am unable to
determine whether there is something wrong in Mockito or the
mock setup. For now, the only course of action is to disable
the tests.

Bug: 70949390
Test: runtest frameworks-telephony
Change-Id: I1ab10f99fd41aa1687993619b16c65bbe10fca74