RequestBuilder for ApduSender
RequestBuilder helps build a list of APDU commands that are sent by
ApduSender. ApduSender accepts a RequestProvider instance to get all
APDU commands after a logical channel is opened. The caller to an
ApduSender can generate different APDU commands based on the select
response of the opened logical channel. This is used by EuiccCard to
provide the implementations for different GSMA RSP specifications.

Bug: 38206971
Test: unit test on the ApduSender CL of the same topic
Change-Id: Icab30ca17a19576a6272a7cb8cc7b8d2cc666003