IMS-VT: Set whether video is enabled locally or not.
For ViLTE calls:

1. UE rejects MT upgrade requests when mobile data is
data is turned OFF and vice-versa.

2. Lower layer indicate that UE is video capable when
mobile data is turned ON and vice-versa.

In issue case, when processing mobile data toggle,
UE is put in state indicating video is not locally
enabled thereby rejecting MT upgrade requests though
lower layers indicated that video is enabled.

To fix this, always inform video provider whether
video is locally enabled or not when updating media
capabilities

Bug: 72143006
Test: Manual
Change-Id: Iccdb5881cef10cedab8d5ff1aad82cc1a1bdb034