Tidy up clock adjustment code / logging
Tidy up some of the clock adjustment code. On a structural
level this removes a few unusual/edge cases that wouldn't
yield particularly useful results:

1) Looking for a UTC-based time zone in a country known not to
use UTC (unlikely to yield a result).
2) Adjusting the system clock when all that is known is the country
(this would previously have set the device to the time zone
the device is using, so probably a no-op).

In the actual adjustment code, rather than adjust the current
device system clock, the time from the NITZ signal is used instead.
The original purpose of this code was apparently to handle NITZ
data where the network is pushing out local time rather than
UTC so it makes sense to base the time off the NITZ data.

This change also makes some non-functional logging changes.

Test: atest FrameworksTelephonyTests
Bug: 63743683
Bug: 37494871
Change-Id: Id0fd20cbe9cf97911ec43eee847aa84b1b32b42f