Add the defaultNetwork parameter to the NetworkStats.Entry.
This allows us to remove a NetworkStats.Entry constructor that is
used only by this code. DEFAULT_NETWORK_YES is used because VT
traffic uses the network specified by the carrier.

While I'm at it, also change "1" to NetworkStats.METERED_YES, so
the code is clearer about what it's doing.

Bug: 35142602
Test: builds, boots
Change-Id: I88a288ed03437c0945b221871919ea0f78a16ebe