EuiccCard #1
EuiccCard extends UiccCard to provide eUICC functionality defined by
GSMA RSP specifications.

This CL is the first CL to migrate the existing EuiccCard code. It
adds the enums, exceptions, empty methods, and the implementation of
getting spec version from the opening logical channel response.
Other implementation of methods are added in following up CLs.

The patch #3 is the existing code and can be compared with the patch #4
to see the actual diff.

Bug: 38206971
Test: unit test
Change-Id: I3b64fd03c35cf8fa154f38423bdb8dad11a309b7