Update patch set 1
Patch Set 1:

> > I can only find jf and klte currently maybe needs this.
 > 
 > At least i9305, herolte and hero2lte as well and probably others.
 > 
 > This is a Samsung specific thing, but device maintainers haven't
 > applied the related patches. This is because the number of users,
 > who require this for MMS to work, is quite small and thus there
 > people are not requesting it. Also people do not really use MMS
 > nowdays.
 > 
 > Here's some context: https://forum.xda-developers.com/galaxy-s3/help-i9305/enable-mms-5-1-aosp-roms-t3183220

 > MMS was working on klte last I checked. On Sprint, anyways. So I
 > don't think we need this hack. (Yes, I'm testing on O). If it does
 > fix a device in any observable way, I'm not opposed to it, I'm just
 > not sure it's needed.

MMS only works if you can access the MMSC via your default APN or wifi. For operators in Finland this is not the case. You need to use a separate APN, which is specifically set-up for accessing only the MMSC. I know there are some operators in Australia as well, which require this.

Patch-set: 1