Fixed network mode loaded issue
SubscriptionInfoUpdater.java adds subId during device power-up and if the SIM card
is new to the slot it would set the network mode as Global. Before assigning network mode
as Global, first check if a network mode already available in DB for the current subId.
If network mode available, use the same as current preferred network mode.

Bug: 36850854
Test: manual
Change-Id: I93e9252a8d7f2f31ea212f04525a7a94b479d1fb