Start OTA in EuiccCardController
We will try to start OTA when we receive the broadcast of SIM slot
status change, since at this time, radio is on and the slot info. is
ready. We will only try to start OTA if we current is on eSIM slot.

Bug: 72880982
Test: Included
Change-Id: Id2fec7dab328f5ccd8b0f0ee7ff83aa24bc1dda3
Merged-In: Id2fec7dab328f5ccd8b0f0ee7ff83aa24bc1dda3