Reduce logspam in carrier privilege status check.
UiccCarrierPrivilegeRules can be called by non-platform code via
system APIs (TelephonyManager#checkCarrierPrivilegesForPackage), and
thus it is not an error worthy of logging if the given package is not
installed. It is also easy to infer that a package is not
installed from a bugreport (or determine when it was installed, if
it is installed later). Just return NO_ACCESS.

Change-Id: If068fadb34ea6256287d9e92f00cd2dd5b5ddc9b
Fixes: 72764153
Test: TreeHugger (CL is trivial)