Update patch set 3
Patch Set 3:

Dunno. Seems weird in general, do CDMA_LTE devices get created as GSM initially because they have a SIM card, and then get set to a different type later? That's the only explanation that makes sense to me as for why this doesn't affect CDMA devices with a SIM card (that also don't use NV, since I guess the telephony stack is trying to create a phone somewhere for NV devices because NV subs don't depend on a SIM card?)

Patch-set: 3