Framework Integration of dynamic query for ImsService
1) Adds support for dynamic query of ImsService features.
When the ImsResolver is notified of a change to the ImsService
package (via PackageManager or CarrierConfig), we perform
an async query to the affected ImsService. The updated
feature configuration is then saved to the ImsServiceInfo
cache and features are recalculated. This will cause the
Service to create/remove features according to this
config. We currently keep the ability for "compat"
ImsServices and device default ImsServices to communicate
their feature configurations via manifest query, but
all carrier ImsServices will need to do a dynamic query
to relay feature configuration information.

2) Fixes issue where we were sometimes not unbinding
to ImsServices, even if there were no features assigned.

3) We now drop down to the device default ImsService
when the SIM card is removed.

Test: Telephony Unit Tests, Telephony Test ImsService
Bug: 72642524

Merged-In: If65f25eb6b0696b38bd55b96c1d5a64d2a112a6b
Change-Id: Ieafa99c1454606016c71e4d183adf445f4189dc2