Replace Manual Parceling for CellInfoCdma
CellInfoCdma has a number of legacy fields with very
strange values. In order to remove some unfortunate
code in CellSignalStrengthCdma that tried to twiddle
the sign on all of the fields during parcel/unparcel
it was necessary (and more than appropriate) to
update RIL to treat the parcel values and ordering
of CellInfoCdma as an implementation detail, instead
calling the respective class constructors and parcel()
methods of the classes.

Bug: 32364031
Test: runtest frameworks-telephony
Change-Id: Id257f8c59476ee083e41dd4161ac0c28759be1e4