Return Unsupported for Keepalive Requests on IWLAN
In order to support NATT Keepalive on IWLAN, we need
to add support to the DataService interface for the
start, stop, and unsol commands. Since this would
be a system API and the plumbing is not trivial, the
safe thing to do is to ensure that requests that come
for a WLAN-backed network are rejected.
The only potential use case for NATT keepalive offload
would be if the remote end of the tunnel terminates
behind a NATT, which is unlikely for IWLAN. Meanwhile
requests for WWAN will funnel directly to Radio, which
is unfortunate, but sufficient for all current use
cases. Thus, while the DataService interface support
should be added as a matter of completeness and
clarity, for now, we will simply reject calls to start
Keepalives on IWLAN.

Bug: 72331356
Test: compilation
Change-Id: I95681594c905c5eb26246a6b3cef4696d587f28e