Make sure CARDSTATE_ABSENT is sent out from UiccSlot.
We need to make sure CARDSTATE_ABSENT is sent from UiccSlot to
SubscriptionInfoUpdater so that SubscriptionInfoUpdater proceeds
necessary steps. We initialize mCardState to be null (as an enum)
instead of CARDSTATE_ABSENT in UiccSlot, so that we can identify
when we first received ABSENT state update, and send out update
to SubscriptionInfoUpdater by then.

Bug: 74608255
Test: unittest
Change-Id: If50594225dc250440eb4a957d05ab78b11e94711
Merged-In: If50594225dc250440eb4a957d05ab78b11e94711