Create UiccCard when mUiccCard is null and mCardState is not ABSENT
Previously mUiccCard is only created when the mCardState is changing
from ABSENT to non ABSENT. This logic is now broken because the
mCardState may be updated in both IccCardStatus and IccSlotStatus, and
may cause the issue that mUiccCard is never created. This Cl added an
extra logic that when the mCardState is not ABSENT, while the mUiccCard
is still null, we also need to create a new mUiccCard.

This is a clean cherry-pick from:
https://googleplex-android-review.git.corp.google.com/c/platform/frameworks/opt/telephony/+/3867960

Bug:77279870
Test: Unit Test
Change-Id: I0750cb118044dca6dfc12e8cf1a9cb6f143b2fdb
Merged-in: I0750cb118044dca6dfc12e8cf1a9cb6f143b2fdb
(cherry picked from commit de72f77373b6b38cd650133c89f6cf72c55ac06b)