Create UiccCard when mUiccCard is null and mCardState is not ABSENT
Previously mUiccCard is only created when the mCardState is changing
from ABSENT to non ABSENT. This logic is now broken because the
mCardState may be updated in both IccCardStatus and IccSlotStatus, and
may cause the issue that mUiccCard is never created. This Cl added an
extra logic that when the mCardState is not ABSENT, while the mUiccCard
is still null, we also need to create a new mUiccCard.

Bug:77279870
Test: Unit Test
Change-Id: I0750cb118044dca6dfc12e8cf1a9cb6f143b2fdb