Mark top flaky test @FlakyTest
This means they will not block presubmit. Individually these tests fail
between 3% and <1% of the time, but since tests run as suites, and any
single test failure will cause a suite failures, the combined effect is
that 9% of all CLs that trigger telephony unit tests will encounter a
failure due to a flake.

Bug: 78291670
Test: tests still build
Change-Id: I85760a29f38c2a6245c06f050c037a92d107ed6a