Add assert to TelephonyTest#setup
SubscriptionController#getInstance has been found to return null during
testing sometimes, causing crashes which are hard to trace. If we can
assert that the method does not return null in setup, it will be easier
to catch the cases where it breaks.

Bug: 78359408
Test: atest FrameworksTelephonyTests
Change-Id: Ie22511b145ccf29824bc8be46be10ebe2dd89970
Merged-In: Ie22511b145ccf29824bc8be46be10ebe2dd89970