Restore the OEM hook implementation and usage
The current plan is to keep the OEM hook implementation in the 
framework, this is for legacy device and old device upgrading 
(per b/78098059).

For P launching device, the OEM hook service is removed in hal, 
so the OEM hook api in framework should return something dummy, and 
prevent infinite loop to get service, but these parts should be done
in seperate CLs after this restore CL is merged.

This reverts commit acc87ce9d9a91f0eb80c411ddbf66c82ef2e04d3.

Bug: 34344851
Change-Id: Ib58d9e0842ac1764f91391b1403a3d841f4a1bd9