Restore the OEM hook implementation and usage
The current plan is to keep the OEM hook implementation in the 
framework, this is for legacy device and old device upgrading 
(per b/78098059).

For P launching device, the OEM hook service is removed in hal, 
so the OEM hook api in framework should return something dummy, and 
prevent infinite loop to get service, but these parts should be done
in seperate CLs after this restore CL is merged.

This reverts commit ae1efe2c57b979420c2e63df88ceef5cefcfe056.

Bug: 34344851
Change-Id: Ieb0b52c433c7f1800edf3e7b739b1c7b55df5790
Test: TH tests; cuttlefish test

Merged-In: Ib58d9e0842ac1764f91391b1403a3d841f4a1bd9