Introduce an interface for NitzStateMachine
Introduce an interface for NitzStateMachine so that a new
implementation can be swapped in easily. This will be part
of the mechanism to revert the time / time zone detection
changes.

A second implementation of NitzStateMachine will be created
in a later change once there is a service for it to communicate
with.

No functional changes are intended by this change.

Bug: 78217059
Test: Build only
Merged-In: Id789312083a3f72b11971b3b00b40b8d60153db4
Change-Id: Id789312083a3f72b11971b3b00b40b8d60153db4
(cherry picked from commit 0ed542c162e5d9c25e051563f412e776989f6a2b)