Report OUT_OF_SERVICE when service is lost but still IMS registered
We currently report IN_SERVICE for voice reg state for IMS when
registered. This is incorrect because we can lose the physical link
to the network when the user moves too far away from the tower,
but IMS will still stay registered for 15 min until timeout.

This change modifies the ServiceState of ImsPhone to report
IN_SERVICE only if we are IMS registered and voice capable
(whether that be VoLTE or VoWiFi).

Bug: 72483562
Test: Manual Callbox tests, atest FrameworksTelephonyTests

Change-Id: I74bda92992fe86bf3a847f67d84aacf12f963020