Move MMTEL feature capabilities reset into MmTelFeatureCompatAdapter.
ImsPhoneCallTracker currently resets the feature capabilities which are
available when IMS is de-registered.  Moving this to the compat adapter
as it is now expected that an ImsService which does not make use of the
compat adapter will inform the framework of capability changes prior to
de-registration.

Bug: 73076695
Test: Enable airplane mode and inspect logs to ensure feature capabilities
are reset to 0.
Merged-In: Iee2ecb3fb9f0a6edf18570fe65abb9827e048fbd
Change-Id: Iee2ecb3fb9f0a6edf18570fe65abb9827e048fbd