Passing data RAT for old setup data call API
The old 1.0 setup data call interface takes RAT as an input
parameter. In order to support the existing legacy devices, we
pass the data RAT instead of unknown to the vendor RIL. This change
does not impact the new setup data API, which takes access network
type as a parameter.

Test: Manual + unit tests
Bug: 79234282
Change-Id: Ibff5913f4ed48006d507281b69e5648412188a45