Share lock object between UiccSlot, UiccCard and UiccProfile.
Prior to P, there was a single class UiccCard that had the lock. Now
it's split between these 3 classes, but if each has a separate lock
there can be a deadlock if they call each other's APIs which they do.
The solution is to use a single lock between them which is equivalent
to how the locking mechanism was earlier.

Test: Basic telephony sanity.
Bug: 80317780
Change-Id: Iadecb726dc3bf85dcde76c70a54e191f1a55007c