Consider subId when reading resource wfcSpnFormat
The resource wfcSpnFormat controls the spn text when Wi-Fi Calling is
registered. wfcSpnFormat often differs depending on mcc/mnc. However
the subId id is not considered when reading wfcSpnFormat, which may
cause an incorrect version of the resource to be read.

Use SubscriptionManager#getResourcesForSubId to get the correct
wfcSpnFormats by considering subId.

Bug: 110432292
Test: Manual
Change-Id: I8e84da5818a0fab9e9a7dcc2cc1d29c4f358c4c9