Create a new implementation of NitzStateMachine
Create a new implementation of NitzStateMachine by copying
OldNitzStateMachine and calling it NewNitzStateMachine.
The associated TimeServiceHelper is also being renamed
OldTimeServiceHelper, and a NewTimeServiceHelper is being
created.

In this initial commit NewNitzStateMachine is a copy of
OldNitzStateMachine with no logic changes.

The old version is kept for now in case we need to switch
back due to a bug and for reference.

Test: build
Test: (internal) atest FrameworksTelephonyTests:com.android.internal.telephony.NewNitzStateMachineTest
Test: (internal) atest FrameworksTelephonyTests:com.android.internal.telephony.OldNitzStateMachineTest
Bug: 78217059

Change-Id: I84d3a3247c6b4982b886e6232ca4c09748eadcb6