IMS: Send call barring request over IMS only if UT is enabled
In the existing Implementation, call barring request is sent
over IMS socket either if IMS is in service or UT is enabled.
So the request is sent over IMS even if UT is not enabled when
IMS is in service.

However modem is trying the request over CS when UT is not enabled
and the request is failing since password is not passed for the
call barring requests sent over IMS.

Test: Manual
Bug: 110187445
Change-Id: I0b9e2bc7dac381ac1d00887757954465c1beddc3