Revert isResolvingImsBinding code in ImsResolver
When querying a carrier ImsService for IMS features
after a new SIM is inserted, there is an indeterminate period
of time where we do not know what features are supported.
During this time telephony gets an independent
CARRIER_CONFIG_CHANGED signal to update the IMS config to the
modem. We were attempting to supress this signal while
determining in the ImsResolver which ImsService MMTEL will
be created on. This resulted in a race condition when the
carrier ImsService only supported the RCS feature because
the carrier config update would be suppressed and the
ImsPhoneCallTracker would never send a config update (because
the device ImsService is still the only MMTEL feature).

Therefore this change reverts the code that detected this
since it was likely that the CARRIER_CONFIG_CHANGED signal
would be processed in other parts of telephony before it
was even processed here.

Bug: 109762920
Test: Telephony GTS/Unit Tests
Merged-In: Ic933566dfbf364feb8f9648bcd95d2d3b0d82772
Change-Id: I20a713b8858931dce2a304fc306bfe8553546f51