Fix incorrect order of arguments / fix bug
android.util.TimestampedValue's constructor arguments
are reversed from the replaced telephony TimeStampedValue
class. This constructor was not reversed when it should
have been.

Also, both values should be adjusted to account for
elapsedRealtimeMillis changing, or neither. The reference
time was not updated but the system clock time was, which
would lead to errors creeping in. This commit changes the
code to use the original NITZ signal values.

Bug: 78217059
Test: Build
Change-Id: I8de75be56624bd737b9312c9655057b0f8251be2