Data activity didn't reflect UDP traffic
Currently, the DataActivity will use the same source with data
stall detection to count packet traffic, which is TCP only.
However, more and more modern applications uses UDP for
communication such as QUIC and streaming. And the data icon
arrows did not reflect the truth that packet is transmitting.
It should be more accurate that data icon arrows should count
all packet traffic.

Note that data stall detection remains the same for now since
we don't want to break functionality for below cases:
1. Out-going UDP packets trigger false positive, such as live
   broadcast apps or b/7903145
2. Incoming local communications cause false negative, such as
   IPv6 Neighbor Discovery messages between modem & AP.

Bug: 110773754
Test: 1. play youtube and see data icon arrows in QuickSetting
      2. enable updateDataActivity debug log and make sure it
         counts.
      2. runtest frameworks-net

Change-Id: Id2f763f396b329ab82af09bb5e084af7201f9d7c