Avoid Crash in NattKeepalive Error Response Case
In the error response case, depending on the error,
a duplicate response message was being sent, first by
the natt keepalive code and then later by the generic
RIL error handling code. This patch reworks the response
error handling code to avoid duplication and changes the
treatment of an UNSUPPORTED response to be treated as
an error both at RIL and at the NattKeepalive tracking
level, which simplifies the logic.

Bug: 110850108
Test: manually verified (hacked) that this patch fixes
    the crash when UNSUPPORTED is sent. Verified by
    test team using callbox.
Change-Id: I34becc4634801b97d2df5afa86462800f555c367