Replace CellLocation with CellIdentity in SST
The SST now gets CellIdentity with every change
in registration state. Rather than constructing a
CellLocation manually from fields in DATA_REG and
VOICE_REG and tracking those, we can directly use
the CellIdentity structures. Then, when needed,
those structures can be converted into CellLocation
objects, which are a strict subset. This reduces
the type of structures, manual conversion, and overall
complexity of tracking the current serving cell in
SST. It also consolidates logic to use the
CellIdentity class, which is far more complete than
the CellLocation. Enhancements:
1) getCellLocation() can now return TD-SCDMA info
2) getCellLocation() will now check that the cell it
   returns actually has identifying information (CID)
3) Remove duplicate logic that clears lat/long from
   CdmaCellLocation (replaced by LocationAccessPolicy)

Bug: 111655742
Bug: 111658462
Bug: 67711865
Test: manual
Change-Id: I8645c0d5d5777d6d0d108e62bc26694f3dc00540