Carrier Signal for secondary uesrs
When a secondary user's profile is logged in and does not have a data plan or has run out of data, the message specified in Secondary User shall be displayed. Applicable to PCO-3, and PCO-5 on devices that support multiple user profiles.
Tapping the notification shall take the user to Settings > Users where they can view their profile.

Test: Go through SetupWizard for secondary user and insert sim during or after SetupWizard is completed.

Change-Id: I2ccce4eacabad4fac73fe8227f110a2df2ff3300
Signed-off-by: Sangcheol Lee <goodsc.lee@samsung.com>