Correct a misunderstanding of the NITZ spec
The NITZ spec doesn't give units for the DST offset
component of the full offset. It turns out the value
can only be for one or two hours which apparently
is taken literally as "number of hours", not specified
in quarter hours like the full offset.

The actual value is not used anywhere currently: we
just check for "was the value provided" and "is value
!= 0". This means that the logic won't change, just
the absolute value / scaling value is being corrected
and logcat will show correct information.

Bug: 112043518
Test: atest com.android.internal.telephony.NitzDataTest
Test: atest com.android.internal.telephony.TimeZoneLookupHelperTest
Change-Id: I1d40225641aa8406090d19cba7d679294067c206